extract multiple git views from index.jsx to its own file, implement a simple tree-view to show changed files in git-commit-view
